fase1/requirements.txt
# FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Image processing
Pillow==10.0.1
piexif==1.1.3
imagehash==4.3.1

# Utilities
pydantic==2.4.2
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.10.1
isort==5.12.0

# Optional: For better development experience
httpx==0.25.2  # For testing API calls